+++ vue-next-admin/src/views/system/role/index.vue
@@ -17,24 +17,39 @@
 				</el-button>
 			</div>
 			<el-table :data="state.tableData.data" v-loading="state.tableData.loading" style="width: 100%">
-				<el-table-column type="index" label="序号" width="60" />
-				<el-table-column prop="roleName" label="角色名称" show-overflow-tooltip></el-table-column>
-				<el-table-column prop="roleSign" label="角色标识" show-overflow-tooltip></el-table-column>
-				<el-table-column prop="sort" label="排序" show-overflow-tooltip></el-table-column>
-				<el-table-column prop="status" label="角色状态" show-overflow-tooltip>
-					<template #default="scope">
-						<el-tag type="success" v-if="scope.row.status">启用</el-tag>
-						<el-tag type="info" v-else>禁用</el-tag>
-					</template>
+				<el-table-column prop="id" label="角色ID" width="130" />
+				<el-table-column prop="name" label="角色名称" show-overflow-tooltip></el-table-column>
+				<el-table-column prop="sequence" label="排序" show-overflow-tooltip></el-table-column>
+        <el-table-column prop="status" label="状态" width="80">
+          <template #default="scope">
+            <el-switch
+                v-model="scope.row.status"
+                class="ml-2"
+                style="--el-switch-on-color: #409eff; --el-switch-off-color: #ff4949"
+                :active-value="1"
+                :inactive-value="2"
+                size="small"
+                inline-prompt
+                active-text="启用"
+                inactive-text="禁用"
+                width="45px"
+                @change="changeStatus(scope.row)"
+            />
+          </template>
+        </el-table-column>
+				<el-table-column prop="memo" label="角色描述" show-overflow-tooltip></el-table-column>
+				<el-table-column prop="created_at" label="创建时间" show-overflow-tooltip>
+					<template #default="scope"> {{ dateStrFormat(scope.row.created_at) }}</template>
 				</el-table-column>
-				<el-table-column prop="describe" label="角色描述" show-overflow-tooltip></el-table-column>
-				<el-table-column prop="createTime" label="创建时间" show-overflow-tooltip></el-table-column>
-				<el-table-column label="操作" width="100">
+				<el-table-column label="操作" width="140">
 					<template #default="scope">
-						<el-button :disabled="scope.row.roleName === '超级管理员'" size="small" text type="primary" @click="onOpenEditRole('edit', scope.row)"
+						<el-button  size="small" text type="primary" @click="onOpenEditRole('edit', scope.row)"
 							>修改</el-button
 						>
-						<el-button :disabled="scope.row.roleName === '超级管理员'" size="small" text type="primary" @click="onRowDel(scope.row)">删除</el-button>
+            <el-button  size="small" text type="primary" @click="onOpenAuthRole(scope.row)"
+            >授权</el-button
+            >
+						<el-button size="small" text type="primary" @click="onRowDel(scope.row)">删除</el-button>
 					</template>
 				</el-table-column>
 			</el-table>
@@ -44,27 +59,32 @@
 				class="mt15"
 				:pager-count="5"
 				:page-sizes="[10, 20, 30]"
-				v-model:current-page="state.tableData.param.pageNum"
+				v-model:current-page="state.tableData.param.page"
 				background
-				v-model:page-size="state.tableData.param.pageSize"
+				v-model:page-size="state.tableData.param.limit"
 				layout="total, sizes, prev, pager, next, jumper"
 				:total="state.tableData.total"
 			>
 			</el-pagination>
 		</div>
 		<RoleDialog ref="roleDialogRef" @refresh="getTableData()" />
+    <RoleAuthDialog  ref="roleAuthDialogRef"/>
 	</div>
 </template>
 
 <script setup lang="ts" name="systemRole">
 import { defineAsyncComponent, reactive, onMounted, ref } from 'vue';
-import { ElMessageBox, ElMessage } from 'element-plus';
+import {ElMessageBox, ElMessage, FormInstance} from 'element-plus';
+import { useRoleApi } from '/@/api/system/role';
+import {SysRoleState,RoleType} from "/@/types/views";
 
 // 引入组件
-const RoleDialog = defineAsyncComponent(() => import('/@/views/system/role/dialog.vue'));
-
+const RoleDialog = defineAsyncComponent(() => import('/@/views/system/role/component/dialog.vue'));
+const RoleAuthDialog = defineAsyncComponent(()=> import('/@/views/system/role/component/authDialog.vue'))
 // 定义变量内容
+const roleApi = useRoleApi();
 const roleDialogRef = ref();
+const roleAuthDialogRef = ref();
 const state = reactive<SysRoleState>({
 	tableData: {
 		data: [],
@@ -72,27 +92,19 @@ const state = reactive<SysRoleState>({
 		loading: false,
 		param: {
 			search: '',
-			pageNum: 1,
-			pageSize: 10,
+			page: 1,
+			limit: 10,
 		},
 	},
 });
 // 初始化表格数据
-const getTableData = () => {
+const getTableData = async () => {
 	state.tableData.loading = true;
-	const data = [];
-	for (let i = 0; i < 20; i++) {
-		data.push({
-			roleName: i === 0 ? '超级管理员' : '普通用户',
-			roleSign: i === 0 ? 'admin' : 'common',
-			describe: `测试角色${i + 1}`,
-			sort: i,
-			status: true,
-			createTime: new Date().toLocaleString(),
-		});
-	}
-	state.tableData.data = data;
-	state.tableData.total = state.tableData.data.length;
+	await roleApi.listRole(state.tableData.param).then((res: any) => {
+		state.tableData.data = res.data.roles;
+		state.tableData.total = res.data.total;
+	});
+	console.log(state.tableData);
 	setTimeout(() => {
 		state.tableData.loading = false;
 	}, 500);
@@ -105,14 +117,40 @@ const onOpenAddRole = (type: string) => {
 const onOpenEditRole = (type: string, row: Object) => {
 	roleDialogRef.value.openDialog(type, row);
 };
+const onOpenAuthRole =(row: object) => {
+  roleAuthDialogRef.value.openAuthDialog(row);
+}
+//修改状态
+const changeStatus = async (obj: any) => {
+  let text = obj.status === 1 ? '启用' : '停用';
+  ElMessageBox({
+    closeOnClickModal: false,
+    closeOnPressEscape: false,
+    title: '警告',
+    message: '确认要 ' + text + ' "' + obj.name + '"角色吗?',
+    showCancelButton: true,
+    confirmButtonText: '确定',
+    cancelButtonText: '取消',
+  })
+      .then(async () => {
+        return await roleApi.updateRoleStatus(obj.id, obj.status);
+      })
+      .then(() => {
+        ElMessage.success(text + '成功');
+      })
+      .catch(() => {
+        obj.status = obj.status === 1 ? 2 : 1;
+      });
+};
 // 删除角色
-const onRowDel = (row: RowRoleType) => {
-	ElMessageBox.confirm(`此操作将永久删除角色名称：“${row.roleName}”，是否继续?`, '提示', {
+const onRowDel = (row: RoleType) => {
+	ElMessageBox.confirm(`此操作将永久删除角色名称：“${row.name}”，是否继续?`, '提示', {
 		confirmButtonText: '确认',
 		cancelButtonText: '取消',
 		type: 'warning',
 	})
-		.then(() => {
+		.then( async () => {
+      await roleApi.deleteRole(row.id)
 			getTableData();
 			ElMessage.success('删除成功');
 		})
@@ -120,12 +158,12 @@ const onRowDel = (row: RowRoleType) => {
 };
 // 分页改变
 const onHandleSizeChange = (val: number) => {
-	state.tableData.param.pageSize = val;
+	state.tableData.param.limit = val;
 	getTableData();
 };
 // 分页改变
 const onHandleCurrentChange = (val: number) => {
-	state.tableData.param.pageNum = val;
+	state.tableData.param.page = val;
 	getTableData();
 };
 // 页面加载时